apt_parser & config: Implement package version download limit
Some repos, such as VS Code's repo, include all prior package versions
in them, but you may not need the prior package versions. This commit
allows for a limited subset of package versions to be downloaded.
It also allows for excluding directories from repo downloads (e.g., you
may not want 'beta' or 'insiders' packages to be mirrored.
Note that these changes are based on parsing the filename, and not any
package metadata. As such, they require that the package names conform
to the proper Debian package naming conventions.