feat: Include snapshot functionality
Allow for hardlink-based snapshots that can be configured via the
configuration file.
Includes the ability to prune snapshots (remove stale ones in a
semi-automated fashion) and the ability to delete individual snapshots.
Snapshots can also be given custom names, though the default is to allow
automated date+time stamps as the snapshot names.
Admins can test and publish individual snapshots, though some of the
logic and handling is still being worked-out.