misc: Namespace change to the gomirror org
Update the Go module's import path from the cybozu-go organization
to the gomirror organization

Yes, this is fork of the cybozu-go/aptutils project, focusing just on
the go-apt-mirror portion. The original application was / is useful, but it was
in low-update maintenance mode, and there were some features that were
missing from it that I wanted. It was also based on some archived, bespoke
dependencies, so I decided to fork it. I hope that the prior maintainers
receive this well. The original project was / is very good, has served
as a good base here, and I'm grateful for their work.